Separated code responsibilities
Various methods/functions were rearranged into their own specific classes for organizational purposes.

Implemented updated library for ExpectedConditions class, which was deprecated
