+++ packages.config
@@ -1,6 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="dotenv.net" version="2.1.1" targetFramework="net471" />
+  <package id="DotNetSeleniumExtras.WaitHelpers" version="3.11.0" targetFramework="net471" />
   <package id="Selenium.Support" version="3.141.0" targetFramework="net471" />
   <package id="Selenium.WebDriver" version="3.141.0" targetFramework="net471" />
   <package id="Selenium.WebDriver.ChromeDriver" version="84.0.4147.3001" targetFramework="net471" />
