+++ src/Others/TODOS.cs
@@ -4,17 +4,9 @@
     {
         /*
         TODO: 
-        - Review & Organize code into different classes
-        - review the use of waitGlobal and driverGlobal, maybe you don't need to pass  'driver' references on methods
-        - Find the updated version of ExpectedConditions as it is deprecated
-        - Modify Web app code to provide feed back when user has already been registered
-        - Modify Web app code to provide feedback when a user tries to login/register with empty strings (?) 
+        - Modify Web app code to provide proper feed back when user has already been registered
         - Assign appropriate return values for methods
-        */
-
-        /*
-        TODO - Tests: 
-        - Think of actual ways to test frontend
+        - Core.cs | line 86 | Find way to check if submitted secret is actually present in list
         */
     }
 }
